[octomap] update octomap for maitaining color point cloud
