+++ eus_pcl/include/eus_pcl/euspcl_octomap.h
@@ -3,7 +3,9 @@
 
 // eus functions
 extern pointer OCTOMAP_CREATE (register context *ctx, int n, pointer *argv);
+extern pointer OCTOMAP_CREATE_COLOR (register context *ctx, int n, pointer *argv);
 extern pointer OCTOMAP_DELETE (register context *ctx, int n, pointer *argv);
+extern pointer OCTOMAP_READ_BINARY (register context *ctx, int n, pointer *argv);
 extern pointer OCTOMAP_READ_NODES (register context *ctx, int n, pointer *argv);
 extern pointer OCTOMAP_READ_UNKNOWN (register context *ctx, int n, pointer *argv);
 extern pointer OCTOMAP_DUMP_DATA (register context *ctx, int n, pointer *argv);
